third_party/blink/web_tests/external/wpt/css/css-properties-values-api/animation/custom-property-animation-length-comma-list-expected.txt

This is a testharness.js-based test.
PASS Animating a custom property of type <length>#
PASS Animating a custom property of type <length># with a single keyframe
PASS Animating a custom property of type <length># with additivity
PASS Animating a custom property of type <length># with a single keyframe and additivity
FAIL Animating a custom property of type <length># with iterationComposite assert_equals: expected "250px, 275px" but got "50px, 75px"
PASS Animating a custom property of type <length># with different lengths is discrete
Harness: the test ran to completion.
